requirements.txt
websocket-client==0.30.0
pyslack==0.1.3
requests==2.6.2
